Repurposed MainActivity as a new login module and moved all previously known MainActivity Functions to a new HomeActivity class.
